made a component for the shop cards to clear up the file
